Applying Api to details view

+++ lib/lib/featutres/cart/presntation/controller/cart_controller.dart
@@ -16,7 +16,7 @@ class CartController extends GetxController {
     quantity--;
 
   }
-  double get totalPrice => cartItems.fold(0, (sum, item) => sum + (item[Constants.price] as double));
+  int get totalPrice => cartItems.fold(0, (sum, item) => sum + (item[Constants.price] as int));
 
 
 }
